chore(nextjs-apollo-client): rename util file
